Fixed Rage Support and Chain Hook not applying movement speed bonuses properly. This was due to an error in buildActiveSkillModList, which did not add 'Condition:CanGainRage' to the global mod list.
This was fixed with a, in my opinion, slightly hacky solution - I match for "Condition" in modifier name to add to the mod list, but it seems to work without affecting other builds (Tested with a couple builds on poe.ninja, no dps or modifier changes on them that I observed).

Side effect of this change is that Withering Steps now grants 'Condition:CanBeElusive' and 'Condition:CanWither' properly.