Show build name in title bar
Display basic information about the current build in the title of the
window. Three flavors exist, depending on the state of the app:

Path of Building
Application is starting or in the build management view (e.g. if you hit
"<< Back" in the UI)

Path of Building - Imported Build (Gladiator)
Build is imported from a Path of Building code, PasteBin, or from a
local XML file. Note that "Imported Build" is literal; Path of Building
does not store the original character name or league in its data.

Path of Building - Unnamed build (CleverName, Gladiator, SSF Delirium)
Build is imported from GGG. The title takes effect when the passive tree
is imported. Importing just the items from a build will have no effect.