Fix for the removable crafts remaining after editing an item.
Normally, a UI control will inherit the shown/hidden status of its anchor. However, if the "collapse" boolean is set to true on the anchor, instead it will inherit its anchor's position when the anchor is hidden. This was erroneously set to true for the displayItemSectionEnchant section, making it and all subsequent sections ignore the shown/hidden property for the overall item editing display.