Rework Malediction
Malediction is a debuff that inflicts 10% increased damage taken and 10% reduced damage dealt.

As of now, it is only implemented for the Occultist node as an enemy 10% increased damage taken mod.
However, The Dark Seer is another source of Malediction and cannot be implemented the same way as we don't want those 10% to stack.

This commit turns Malediction into a flag and adds support for The Dark Seer's mod that inflicts it.

[v2] use modDB rather than enemyDB in doActorMisc(env, actor)