Simulation Trigger Enhancements (#2456)
* More accurately simulates triggered attacks

Primary bug fix: cooldowns (attack or triggers) now start counting from
when they actually happen instead of from when they last came off
cooldown.
Decoupled the trigger (attack) from the server tick as per
https://www.reddit.com/r/pathofexile/comments/ll1y91/114_hits_flicker_strike_build/gnqj242/?context=10
This better represents almost all triggers such as CoC with a channeling
skill (cyclone) or multistrike, CWDT, CwS, CoK, and CoD. Notably it can
now be incorrect for CoC on 1 attack per press skill at high attack
speeds might need to be special cased if it becomes an issue.

As it can now handle trigger events that happen in the middle of a tick
we enter uncharted waters, we know the cd can only refresh on a server
tick but when does it start. Either at the beginning of the tick that
included the trigger, timestamped at trigger time and prorated come the
tick, or at the start of the next tick. I think we can rule out start of
the next tick as that would lead to very noticable lag. I choose to
model start of the current tick because the numbers end up nicest
(namely preserves the general knowledge that ICDR is primarily a
breakpoint stat). If the cooldown starts immediately (timestamped)
instead small amounts of ICDR value vastly fluctuates depending on how well attack
speed is synced with the server tick (in general this model leads to a
lot of missed triggers and an annoying floating point precision issue
so I wanted to avoid it unless proven true). Will try to get in contact
with GGG for insight into this issue.

* Minor formatting and Spelling Fix

* also use in 1-spell setup

Co-authored-by: Moneypouch <zap911@msn.com>