Fix Phantasmal Flameblast AoE values (#6813)
Phantasmal Flameblast grants base radius to the gem as its quality effect. It used the `active_skill_base_radius_+` stat to grant the extra base radius which was stat mapped to `skill("radiusExtra", nil)`.
This would be overwritten by the stat map on the gem that also used `radiusExtra` instead of summing the values together. Changed the statMap on the gem to use `mod("AreaOfEffect", "BASE")` instead so it sums correctly

Co-authored-by: LocalIdentity <localidentity2@gmail.com>